removing unnecessary models and implementing logic for addign new countries

+++ MRP/Controllers/DrzavasController.cs
@@ -4,9 +4,11 @@ using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
+using System.Net.Http;
 using System.Web;
 using System.Web.Mvc;
 using MRP.Models;
+using Newtonsoft.Json;
 
 namespace MRP.Controllers
 {
@@ -17,13 +19,12 @@ namespace MRP.Controllers
         // GET: Drzavas
         public ActionResult Index()
         {
-            return View(db.Drzavas.ToList());
+            return View(db.Drzavas.OrderBy(x=>x.Naziv).ToList());
         }
 
         // GET: Drzavas/Details/5
         public ActionResult Details(long? id)
         {
-           
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,7 +40,15 @@ namespace MRP.Controllers
         // GET: Drzavas/Create
         public ActionResult Create()
         {
-            return View();
+            // Get all countries from external api
+            HttpClient client = new HttpClient();
+            HttpResponseMessage response = client.GetAsync("https://restcountries.eu/rest/v2/all").Result;
+            response.EnsureSuccessStatusCode();
+            var result = response.Content.ReadAsStringAsync().Result;
+            List<Country> countries = JsonConvert.DeserializeObject<List<Country>>(result);
+            Drzava drzava = new Drzava();
+            drzava.DrzavaDropDown = countries.Select(x => x.Name).ToList();
+            return View(drzava);
         }
 
         // POST: Drzavas/Create
@@ -51,6 +60,22 @@ namespace MRP.Controllers
         {
             if (ModelState.IsValid)
             {
+                // Check name
+                if (string.IsNullOrEmpty(drzava.Naziv))
+                {
+                    ModelState.AddModelError("Naziv", "Ime drzave je obavezno polje.");
+                }
+                else
+                {
+                    if (db.Drzavas.Any(x => x.Naziv.Equals(drzava.Naziv)))
+                    {
+                        ModelState.AddModelError("Naziv", "Drzava sa unetim nazivom već postoji.");
+                    }
+                }
+                if (!ModelState.IsValid)
+                {
+                    return View(drzava);
+                }
                 db.Drzavas.Add(drzava);
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -67,6 +92,13 @@ namespace MRP.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             Drzava drzava = db.Drzavas.Find(id);
+            // Get all countries from external api
+            HttpClient client = new HttpClient();
+            HttpResponseMessage response = client.GetAsync("https://restcountries.eu/rest/v2/all").Result;
+            response.EnsureSuccessStatusCode();
+            var result = response.Content.ReadAsStringAsync().Result;
+            List<Country> countries = JsonConvert.DeserializeObject<List<Country>>(result);
+            drzava.DrzavaDropDown = countries.Select(x => x.Name).ToList();
             if (drzava == null)
             {
                 return HttpNotFound();
@@ -83,6 +115,22 @@ namespace MRP.Controllers
         {
             if (ModelState.IsValid)
             {
+                // Check name
+                if (string.IsNullOrEmpty(drzava.Naziv))
+                {
+                    ModelState.AddModelError("Naziv", "Ime drzave je obavezno polje.");
+                }
+                else
+                {
+                    if (db.Drzavas.Any(x => x.Id != drzava.Id && x.Naziv.Equals(drzava.Naziv)))
+                    {
+                        ModelState.AddModelError("Naziv", "Drzava sa unetim nazivom već postoji.");
+                    }
+                }
+                if (!ModelState.IsValid)
+                {
+                    return View(drzava);
+                }
                 db.Entry(drzava).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -111,6 +159,12 @@ namespace MRP.Controllers
         public ActionResult DeleteConfirmed(long id)
         {
             Drzava drzava = db.Drzavas.Find(id);
+            if (db.Dobavljacs.Any(x=> x.Drzava == id) || db.Zaposlenis.Any(x=> x.Drzava == id))
+            {
+                ModelState.AddModelError("Naziv", "Drzava u upotrebi.");
+                return View(drzava);
+            }
+
             db.Drzavas.Remove(drzava);
             db.SaveChanges();
             return RedirectToAction("Index");
